+++ documentation/docs/src/lib/components/Form/DatePicker/DatePicker.stories.svelte
@@ -12,21 +12,10 @@
 ## 🚀 Installation
 
   \`\`\`bash
- pnpm add @internationalized/date @flightlesslabs/dodo-ui-date
+ pnpm add bits-ui @internationalized/date @flightlesslabs/dodo-ui-date
   \`\`\`
 
-## 🎨 Add Styles
-
- Import global styles in your root component. use the root layout (+layout.svelte) for SvelteKit.
-
- \`\`\`ts
- import '@flightlesslabs/dodo-ui-date/styles/main.css';
- \`\`\`
-
- or scss
- \`\`\`ts
-import '@flightlesslabs/dodo-ui-date/styles/main.scss';
- \`\`\`
+For full setup (styles, themes, configuration), see the [Getting Started](https://github.com/flightlesslabs/dodo-ui/tree/main/packages/dodo-ui-date) guide.
 
 ## ✨ Usage