+++ packages/dodo-ui-bits/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flightlesslabs/dodo-ui-bits
 
+## 0.1.2
+
+### Patch Changes
+
+- Select clerable option added. storybook support added
+
 ## 0.1.1
 
 ### Patch Changes